feat: sideNav.tsx change to SideNav.vue
