+++ src/layouts/components/SideNav.vue
@@ -1,0 +1,109 @@
+<template>
+  <div :class="sideNavCls">
+    <t-menu :class="menuCls" :theme="theme" :value="active" :collapsed="collapsed" :default-expanded="defaultExpanded">
+      <template #logo>
+        <span v-if="showLogo" :class="`${prefix}-side-nav-logo-wrapper`" @click="goHome">
+          <component :is="getLogo" :class="`${prefix}-side-nav-logo-${collapsed ? 't' : 'tdesign'}-logo`" />
+        </span>
+      </template>
+      <menu-content :nav-data="menu" />
+      <template #operations>
+        <span class="version-container"> {{ !collapsed && 'TDesign Starter' }} {{ pgk.version }} </span>
+      </template>
+    </t-menu>
+    <div :class="`${prefix}-side-nav-placeholder${collapsed ? '-hidden' : ''}`"></div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { computed, onMounted } from 'vue';
+import { useRouter } from 'vue-router';
+import union from 'lodash/union';
+
+import { useSettingStore } from '@/store';
+import { prefix } from '@/config/global';
+import pgk from '../../../package.json';
+import { MenuRoute } from '@/types/interface';
+import { getActive, getRoutesExpanded } from '@/router';
+
+import AssetLogo from '@/assets/assets-t-logo.svg?component';
+import AssetLogoFull from '@/assets/assets-logo-full.svg?component';
+import MenuContent from './MenuContent.vue';
+
+const MIN_POINT = 992 - 1;
+const props = withDefaults(
+  defineProps<{
+    menu: MenuRoute[];
+    showLogo: boolean;
+    isFixed: boolean;
+    layout: string;
+    headerHeight: string;
+    theme: string;
+    isCompact: boolean;
+  }>(),
+  { menu: () => [], showLogo: true, isFixed: true, layout: '', headerHeight: '64px', theme: 'light', isCompact: false },
+);
+
+const collapsed = computed(() => useSettingStore().isSidebarCompact);
+
+const active = computed(() => getActive());
+
+const defaultExpanded = computed(() => {
+  const path = getActive();
+  const parentPath = path.substring(0, path.lastIndexOf('/'));
+  const expanded = getRoutesExpanded();
+  return union(expanded, parentPath === '' ? [] : [parentPath]);
+});
+
+const sideNavCls = computed(() => {
+  const { isCompact } = props;
+  return [
+    `${prefix}-sidebar-layout`,
+    {
+      [`${prefix}-sidebar-compact`]: isCompact,
+    },
+  ];
+});
+
+const menuCls = computed(() => {
+  const { showLogo, isFixed, layout } = props;
+  return [
+    `${prefix}-side-nav`,
+    {
+      [`${prefix}-side-nav-no-logo`]: !showLogo,
+      [`${prefix}-side-nav-no-fixed`]: !isFixed,
+      [`${prefix}-side-nav-mix-fixed`]: layout === 'mix' && isFixed,
+    },
+  ];
+});
+
+const router = useRouter();
+const settingStore = useSettingStore();
+
+const autoCollapsed = () => {
+  const isCompact = window.innerWidth <= MIN_POINT;
+  settingStore.updateConfig({
+    isSidebarCompact: isCompact,
+  });
+};
+
+onMounted(() => {
+  autoCollapsed();
+  window.onresize = () => {
+    autoCollapsed();
+  };
+});
+
+const goHome = () => {
+  router.push('/dashboard/base');
+};
+
+const getLogo = () => {
+  if (collapsed.value) {
+    return AssetLogo;
+  }
+  return AssetLogoFull;
+};
+</script>
+
+<style lang="less" scoped></style>
