+++ src/layouts/index.tsx
@@ -6,7 +6,7 @@ import { usePermissionStore, useSettingStore, useTabsRouterStore } from '@/store
 import LayoutHeader from './components/Header.vue';
 import LayoutBreadcrumb from './components/Breadcrumb.vue';
 import LayoutFooter from './components/Footer.vue';
-import LayoutSideNav from './components/SideNav';
+import LayoutSideNav from './components/SideNav.vue';
 import LayoutContent from './components/Content.vue';
 import Setting from './setting.vue';
